+++ CHANGELOG.md
@@ -21,7 +21,9 @@ NOTE: please use them in this order.
 
 ### Bugfixes
 
-- Stripped `include` directives to prevent `AttributeError` in Sphinx ([#114](https://github.com/rstcheck/rstcheck-core/pull/114))
+- Stripped `include` directives to prevent `AttributeError` in Sphinx. Included files
+  are checked separately for existence and not just ignored.
+  ([#114](https://github.com/rstcheck/rstcheck-core/pull/114))
 
 ## [v1.2.2 (2025-06-01)](https://github.com/rstcheck/rstcheck-core/releases/v1.2.2)